revise unit test
